Add files via upload
Arduino controller for lizard vivarium. Handles extremes of cold and heat. One relay is connected to a supplementary heater that should come on when it is too cold, and the other relay is connected to a UV strip light, that should go off it is too hot. Normally you would have a seperate thermostatic spot light to prove the main heat, and a separate timer to turn off the UV strip light and spot at night. In cold conditions it is then likely the Arduino would turn on the supplementary heater. In very hot conditions it is likely the Arduino would turn off the strip light, even though it is timed to be on. Both are safety measures to prevent extremes of heat and cold.